+++ analysis/import/SourceCodeParser/src/main/kotlin/de/maibornwolff/codecharta/importer/sourcecodeparser/SourceCodeParserMain.kt
@@ -7,6 +7,7 @@ import de.maibornwolff.codecharta.serialization.OutputFileHandler
 import de.maibornwolff.codecharta.serialization.ProjectDeserializer
 import de.maibornwolff.codecharta.tools.interactiveparser.InteractiveParser
 import de.maibornwolff.codecharta.tools.interactiveparser.ParserDialogInterface
+import de.maibornwolff.codecharta.tools.interactiveparser.util.InteractiveParserHelper
 import picocli.CommandLine
 import java.io.BufferedWriter
 import java.io.File
@@ -21,9 +22,9 @@ import java.nio.file.Paths
 import java.util.concurrent.Callable
 
 @CommandLine.Command(
-    name = "sourcecodeparser",
-    description = ["generates cc.json from source code"],
-    footer = ["This program uses the SonarJava, which is licensed under the GNU Lesser General Public Library, version 3.\nCopyright(c) 2020, MaibornWolff GmbH"]
+    name = InteractiveParserHelper.SourceCodeParserConstants.name,
+    description = [InteractiveParserHelper.SourceCodeParserConstants.description],
+    footer = [InteractiveParserHelper.SourceCodeParserConstants.footer]
 )
 class SourceCodeParserMain(
     private val outputStream: PrintStream,
@@ -131,4 +132,11 @@ class SourceCodeParserMain(
     }
 
     override fun getDialog(): ParserDialogInterface = ParserDialog
+    override fun isApplicable(resourceToBeParsed: String): Boolean {
+        return false
+    }
+
+    override fun getName(): String {
+        return InteractiveParserHelper.SourceCodeParserConstants.name
+    }
 }